refactor(lld/DeviceIllustration): extract component to its own file
